--- AzureDevopsStateTracker/Entities/Entity.cs
@@ -1,22 +1,0 @@
-﻿using System;
-
-namespace AzureDevopsStateTracker.Entities
-{
-    public abstract class Entity
-    {
-        public string Id { get; protected set; }
-        public DateTime CreatedAt { get; private set; }
-
-        public Entity(string id)
-        {
-            Id = id;
-            CreatedAt = DateTime.UtcNow;
-        }
-
-        public Entity()
-        {
-            Id = Guid.NewGuid().ToString();
-            CreatedAt = DateTime.UtcNow;
-        }
-    }
-}
