+++ AzureDevopsTracker/AzureDevopsTracker.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.31402.337
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AzureDevopsTracker", "AzureDevopsTracker.csproj", "{26FE19B9-7CD6-4418-9112-2E3C363985BF}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{26FE19B9-7CD6-4418-9112-2E3C363985BF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{26FE19B9-7CD6-4418-9112-2E3C363985BF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{26FE19B9-7CD6-4418-9112-2E3C363985BF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{26FE19B9-7CD6-4418-9112-2E3C363985BF}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {53133A2B-5357-47B5-A1C4-30953F789623}
+	EndGlobalSection
+EndGlobal
